fix size_t

+++ examples/functor.cpp
@@ -10,7 +10,7 @@ using namespace proton;
 
 /** define a function interface type.
  */
-typedef long unsigned int (_str2int)(const char*);
+typedef size_t (_str2int)(const char*);
 
 /** define a functor type.
  */
@@ -23,7 +23,7 @@ typedef fp_<_str2int> str2int_fp;
 /** an function object type inheriting the interface.
  */
 struct _f1: fo_<_str2int>{
-    long unsigned int operator()(const char* s)
+    size_t operator()(const char* s)
     {
         return string(s).size();
     }
@@ -35,7 +35,7 @@ typedef ref_<_f1> str2int_fo;
 struct _t1{
     /** the candidate member function.
      */
-    long unsigned int f(const char* s)
+    size_t f(const char* s)
     {
         return strlen(s);
     }
